* ImportRethinkDiffAction update

+++ application/FSS.Omnius.Modules/Tapestry/Actions/Entitron/ImportRethinkDiffAction.cs
@@ -21,7 +21,7 @@ namespace FSS.Omnius.Modules.Tapestry.Actions.Entitron
     {
         public override int Id => 1099;
 
-        public override string[] InputVar => new string[] { "v$Diff", "?s$TablesNames", "?s$ExtIdColumnName", "?s$DiffIdColumnName" };
+        public override string[] InputVar => new string[] { "v$Diff", "?s$TablesNames", "?s$ExtIdColumnName", "?s$DiffIdColumnName", "?s$IsDeletedColumnName", "?b$AllowPermanentDelete" };
 
         public override string Name => "Import RethinkDB diff";
 
@@ -33,6 +33,8 @@ namespace FSS.Omnius.Modules.Tapestry.Actions.Entitron
         private DBTable table;
         private string extIdColumnName;
         private string diffIdColumnName;
+        private string isDeletedColumnName;
+        private bool allowPermanentDelete;
 
         public override void InnerRun(Dictionary<string, object> vars, Dictionary<string, object> outputVars, Dictionary<string, object> InvertedInputVars, Message message)
         {
@@ -43,6 +45,8 @@ namespace FSS.Omnius.Modules.Tapestry.Actions.Entitron
             string tableNames = vars.ContainsKey("TablesNames") ? (string)vars["TablesNames"] : "";
             extIdColumnName = vars.ContainsKey("ExtIdColumnName") ? (string)vars["ExtIdColumnName"] : "ext_id";
             diffIdColumnName = vars.ContainsKey("DiffIdColumnName") ? (string)vars["DiffIdColumnName"] : "id";
+            isDeletedColumnName = vars.ContainsKey("IsDeletedColumnName") ? (string)vars["IsDeletedColumnName"] : "";
+            allowPermanentDelete = vars.ContainsKey("AllowPermanentDelete") ? (bool)vars["AllowPermanentDelete"] : false;
 
             if (diffObject == null)
                 throw new Exception($"{Name}: Diff must not be null");
@@ -91,7 +95,7 @@ namespace FSS.Omnius.Modules.Tapestry.Actions.Entitron
                 }
             }
             catch (Exception e) {
-                throw;
+                throw new Exception($"{Name}: Fatal error occured ({e.Message})");
             }
         }
 
@@ -104,17 +108,20 @@ namespace FSS.Omnius.Modules.Tapestry.Actions.Entitron
 
         private void SetDBItemValues(ref DBItem item, IEnumerable<DBColumn> columns, JToken values)
         {
+            Dictionary<string, object> parsedValues = new Dictionary<string, object>();
+            TapestryUtils.ParseJObjectRecursively((JObject)values, parsedValues);
+
             foreach (DBColumn column in columns) 
             {
                 if (column.Name == DBCommandSet.PrimaryKey) {
                     continue;
                 }
 
-                if (values[column.Name] != null) {
-                    item[column.Name] = values[column.Name].ToObject<object>();
+                if (parsedValues.ContainsKey(column.Name) &&  parsedValues[column.Name] != null) {
+                    item[column.Name] = parsedValues[column.Name];
                 }
                 if (column.Name == extIdColumnName) {
-                    item[column.Name] = values[diffIdColumnName].ToObject<object>();
+                    item[column.Name] = parsedValues[diffIdColumnName];
                 }
             }
         }
@@ -146,7 +153,13 @@ namespace FSS.Omnius.Modules.Tapestry.Actions.Entitron
             var item = table.Select().Where(t => t.Column(extIdColumnName).Equal(id)).SingleOrDefault();
 
             if(item != null) {
-                table.Delete(item);
+                if (isDeletedColumnName != "" && table.Columns.Any(c => c.Name == isDeletedColumnName)) {
+                    item[isDeletedColumnName] = true;
+                    table.Update(item, (int)item[DBCommandSet.PrimaryKey]);
+                }
+                else if(allowPermanentDelete) {
+                    table.Delete(item);
+                }
             }
         }
     }
